+++ src/sha1.h
@@ -43,8 +43,8 @@ class SHA1
 		/*
 		 * Provide input to SHA1
 		 */
-		void Input(const unsigned char* message_array, unsigned length);
-		void Input(const char* message_array, unsigned length);
+		void Input(const unsigned char* message_array, size_t length);
+		void Input(const char* message_array, size_t length);
 		void Input(unsigned char message_element);
 		void Input(char message_element);
 		SHA1& operator<<(const char* message_array);
